requirements.txt
pyyaml>=6.0
paramiko>=2.12.0
rich>=13.0.0
